move environments to use IORef underneath
Enable recursive bindings (and closures) to work properly by nesting
environments within an IORef, to avoid problems with recursive lookup
when the maps are immutable.

At this point it's no longer clear if the state monad is needed other
than for the sake of convenience, it is a bit misleading.